requirements.txt
fastapi==0.115.1
uvicorn==0.34.0
langchain==0.3.25
langchain-core==0.3.25
langchain-google-genai==2.1.4
google-ai-generativelanguage==0.6.18
pydantic==2.10.3
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.2